Refactor editors to /components. Add preview code to editors
