+++ components/PlaintextEditor/index.js
@@ -1,14 +1,20 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import css from './style.css';
 
 function PlaintextEditor({ file, write }) {
-  console.log(file, write);
+  const [value, setValue] = useState('');
+
+  useEffect(() => {
+    (async () => {
+      setValue(await file.text());
+    })();
+  }, [file]);
+
   return (
     <div className={css.editor}>
-      <h3>TODO</h3>
-      <i>text/plain</i>
+      {value}
     </div>
   );
 }
